Tracked down the FOV bug in the ray trace calculation, as per problem identified by MH.  THe first problem was poorly accounting for zero padding in the main path.  If that wasn't enough, the second problem was I had put a zero pad in the rtGeometry call as well, and that wasn't accounted for.  Ugh.
I have zero'd out the rtGeometry padding.  If it is not needed, the whole variable in that file should be removed.

I have corrected the original zero padding problem by making the correct call to oiPad, which is what was properly done in the diffraction and shift-invariant cases.

The scripts still run.  And there is a new validation script, v_oiPad. That tests whether oiPad preserves the spatial resolution as we add more columns to an image.  That is passing validation now, and I will continue to check for a while.