ieWebGet ...
I simplified a lot, possibly removing a feature that Dave Cardinal uses ('read', rather than 'fetch').  The goal is to make the system easier to maintain.

Also, I moved the 'hdr' files into wandell-cardinal and I merged the multispectral and hyperspectral files all into spectral.  As they are loaded here, we can't really tell the difference between them anyway.

The documentation for each type is on the SDR page.