Closing in on the problem ...
The wvf 2 oi is working pretty well for the diffraction limited case.

But the WVF side, including wvfPlot and PSF, seems to be right only for the human 17mm focal length, but not for other focal lengths.

So, we can build wvf and convert to oi, where that should be OK. In the diffraction limited case.