Validation clean ups and script improvements
TODO: Eliminate dctMatrix and otf write.  Not sure where that is coming from.

I fixed the s_humanDisplayPSF issue.  It had to do with showing the small CFA image in the sensor window when the sensor was a human cone mosaic.  The CFA is very large, and we were scaling it by 32.  So that was too large.

I made the tutorial live script the default, eliminating the mfile with the same name.