Eliminate some warnings caused by treating vector as a scalar
In 2024 pre-release, a warning is thrown if you use a vector in a colon expression.  It then just uses the first element of the vector.

A typical case where this was throwing a warning as in

rows = 1:sz

with sz a vector of row/col dimensions.  And sometimes sz arrives as a scalar and sometimes as a vector.

I fixed these up as best I could.