Remove wvfCreate, wvfSet, and wvfGet options for fields that are no longer used in wvfComputePupilFunction
Confusing to allow them, as they now mysteriously have no effect. After tracking that down, I deleted them.  Currently we throw and error if someone tries to set or get.

Also added comment to the tutorial that as written, the passed defocus is interpretted a microns, as someone might think it was diopeters.