Changes to prevent a wvf from being stuck directly into an oi.
1) It is now an error to call
	oi = oiSet(oi,'wvf', wvf);
If you do, you will get an error message telling you to change
it to
	oi = oiSet(oi,'optics wvf', wvf);
which will help make it clear in calling code where the wvf actually lives.

2) We currently continue to allow
	oi = oiSet(oi,'wvf paramname',wvfparamval);
but this now does the set on the wvf that is in the optics structure.
It is an error to try to set if there is no wvf in the optics structure.

We do not exectute a wvfCompute on the updated wvf. Nor do we enforce
that the value set into the wvf is consistent with any redundant values
in the oi or other parts of the optics. So, any user who
is mucking with a wvf that is buried in an oi better know what they are doing.

It is possible we should simply not allow these sets.

3) We also refrained from implementing
	opticsSet(optics,'wvf param',val)
	opticsGet(optics,'wvf param')
as synonyms for getting the wvf, setting one of its params, and then
setting that back into the optics.  We think we want to limit how
much we encourage anyone to think of the wvf attached to the oi/optics
as a dynamic thing.  (See 2 just above.)

4) We looked for but did not find any actual calls to oiSet(...,'wvf',...)