Comments and clean up around the image processor
The data.results field is now clearly identified as containing the linear primary intensities for the model display in the ip.

Calculating sRGB values, based on these intensities and the display model, is managed by displayRender and now more clearly commented.  We show sRGB on the user's display, even if they are modeling a different type of display.

The comments in various