Many improvements to split pixel modeling
Added the spectral QE of the OVT design
Made the design run with the parameters from the paper more closely.
Created the OVT model with 3, not 4 images.
Got rid of the black level of the IMX490, which was causing problems in dark regions because everything was very small.
Many comments.